+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
+## [1.2] - 2021-08-21
+### Added
+- `JSONFunctions`: add `append2Digits` and `append3Digits`
+- `JSONFunctions`: add `displayString`
+
 ## [1.1] - 2021-08-14
 ### Added
 - `JSONFunctions`: split `appendChar` out into separate function